rename old batch script and write boilerplate
