+++ examples/multi_blade_row_example.py
@@ -16,7 +16,7 @@ This basic example shows how to set up a multi blade row meshing instance and ex
 
 import os
 
-from ansys.turbogrid.core.multi_blade_row.multi_blade_row import MultiBladeRow
+from ansys.turbogrid.core.multi_blade_row.multi_blade_row_batch import MultiBladeRow
 
 #################################################################################
 # Create and use a MultiBladeRow instance
